updated all things

+++ Meridian_Web/Meridian_Web/Areas/Client/ViewModels/Price/PriceViewModel.cs
@@ -12,7 +12,7 @@
             
         }
 
-        public int? StartPrice { get; set; }
-        public int? EndPrice { get; set; } 
+        public decimal? StartPrice { get; set; }
+        public decimal? EndPrice { get; set; } 
     }
 }
